--- app/Providers/AuthServiceProvider.php
@@ -1,28 +1,0 @@
-<?php
-/**
- * NOTICE OF LICENSE.
- *
- * UNIT3D Community Edition is open-sourced software licensed under the GNU Affero General Public License v3.0
- * The details is bundled with this project in the file LICENSE.txt.
- *
- * @project    UNIT3D Community Edition
- *
- * @author     HDVinnie <hdinnovations@protonmail.com>
- * @license    https://www.gnu.org/licenses/agpl-3.0.en.html/ GNU Affero General Public License v3.0
- */
-
-namespace App\Providers;
-
-use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Auth;
-
-class AuthServiceProvider extends ServiceProvider
-{
-    /**
-     * Register any authentication / authorization services.
-     */
-    public function boot(): void
-    {
-        Auth::provider('cache-user', fn () => resolve(CacheUserProvider::class));
-    }
-}
