+++ bootstrap/app.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Providers\AppServiceProvider;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
@@ -15,7 +16,7 @@ return Application::configure(basePath: dirname(__DIR__))
     )
     ->withMiddleware(function (Middleware $middleware) {
         $middleware->redirectGuestsTo(fn () => route('login'));
-        $middleware->redirectUsersTo(RouteServiceProvider::HOME);
+        $middleware->redirectUsersTo(AppServiceProvider::HOME);
 
         $middleware->throttleWithRedis();
